+++ utils/shadcn-docs/index.ts
@@ -1,7 +1,7 @@
 import * as Avatar from "./avatar";
 import * as Button from "./button";
 import * as Card from "./card";
-import * as Checkbox from "./checkbox";
+// import * as Checkbox from "./checkbox";
 import * as Input from "./input";
 import * as Label from "./label";
 import * as RadioGroup from "./radio-group";
@@ -12,7 +12,7 @@ const shadcnDocs = [
   Avatar,
   Button,
   Card,
-  Checkbox,
+  // Checkbox,
   Input,
   Label,
   RadioGroup,
