--- utils/shadcn-v2.ts
@@ -1,29 +1,0 @@
-import {
-  avatar,
-  badge,
-  button,
-  card,
-  checkbox,
-  input,
-  label,
-  radioGroup,
-  select,
-  textarea,
-  utils,
-} from "./shadcn-v1";
-
-export const essentialShadcnComponents = {
-  utils: utils,
-  avatar: avatar,
-  // badge: badge,
-  button: button,
-  card: card,
-  checkbox: checkbox,
-  input: input,
-  label: label,
-  radioGroup: radioGroup,
-  select: select,
-  textarea: textarea,
-  // dialog: dialog,
-  // table: table,
-};
