+++ utils/shadcn.ts
@@ -3080,3 +3080,19 @@ export const shadcnComponents = {
   tooltip: tooltip,
   useToast: useToast,
 };
+
+export const essentialShadcnComponents = {
+  utils: utils,
+  avatar: avatar,
+  button: button,
+  card: card,
+  checkbox: checkbox,
+  input: input,
+  label: label,
+  radioGroup: radioGroup,
+  select: select,
+  textarea: textarea,
+  // badge: badge,
+  // dialog: dialog,
+  // table: table,
+};
